+++ src/main/java/org/grnet/pidmr/service/AdminService.java
@@ -4,9 +4,19 @@ import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.core.UriInfo;
 import org.grnet.pidmr.dto.RoleChangeRequestDto;
+import org.grnet.pidmr.dto.ValidatorResponse;
+import org.grnet.pidmr.enums.Validator;
 import org.grnet.pidmr.mapper.UsersRoleChangeRequestMapper;
+import org.grnet.pidmr.pagination.Page;
 import org.grnet.pidmr.pagination.PageResource;
+import org.grnet.pidmr.pagination.PageableImpl;
 import org.grnet.pidmr.repository.RoleChangeRequestsRepository;
+import org.grnet.pidmr.util.Utility;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.stream.Collectors;
 
 @ApplicationScoped
 public class AdminService {
@@ -26,5 +36,26 @@ public class AdminService {
         // Retrieve pageable data from the repository
         var pageable = roleChangeRequestsRepository.fetchRoleChangeRequestByPage(page, size);
         return new PageResource<>(pageable, UsersRoleChangeRequestMapper.INSTANCE.roleChangeRequestsToDto(pageable.list()), uriInfo);
+    }
+
+    public PageResource<ValidatorResponse> getValidatorsByPage(int page, int size, UriInfo uriInfo){
+
+        var allValidators =  Arrays.stream(Validator.values())
+                .map(validator -> new ValidatorResponse(validator.name(), validator.getDescription()))
+                .collect(Collectors.toList());
+
+        var partition = Utility.partition(new ArrayList<>(allValidators), size);
+
+        var validators = partition.get(page) == null ? Collections.EMPTY_LIST : partition.get(page);
+
+        var pageable = new PageableImpl<ValidatorResponse>();
+
+        pageable.list = validators;
+        pageable.index = page;
+        pageable.size = size;
+        pageable.count = allValidators.size();
+        pageable.page = Page.of(page, size);
+
+        return new PageResource<>(pageable, uriInfo);
     }
 }