+++ src/main/resources/db/migration/V25.0__validator_column.sql
@@ -1,0 +1,11 @@
+-- ------------------------------------------------
+-- Version: v25.0
+--
+-- Description: Migration that introduces a new Provider column
+-- -------------------------------------------------
+
+ALTER TABLE Provider ADD COLUMN validator smallint DEFAULT 0;
+
+UPDATE Provider SET validator = 0;
+
+